% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{abbr2state}
\alias{abbr2state}
\title{Make sure all states are there, rename states to regions for plotting maps}
\usage{
abbr2state(state_str)
}
\arguments{
\item{state_str}{State two letter codes as strings}
}
\value{
Returns the state names as full names
}
\description{
Make sure all states are there, rename states to regions for plotting maps
}
\examples{
abbr2state(c("AK", "AL", "IA"))
}
