% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add_data}
\alias{add_data}
\title{Update table data based on string matches}
\usage{
add_data(tre, dat, by, fun = paste0, fun.dat = fun, fun.tre = fun)
}
\arguments{
\item{tre}{treedata object}

\item{dat}{data.frame containing the column \code{by}.}

\item{by}{character vector of names from the \code{dat} that are passed to \code{fun.dat}.
matches to substrings of 0 or 1 node or tip labels in the input tree.}

\item{fun}{function for transforming tree names and data columns. The
default is \code{paste0}. For the tree, this default simply returns the
original names (i.e., is equivalent to the identity function). For
\code{dat}, which is a data.frame, this will paste together elements from
each column specified in \code{by}.}

\item{fun.dat}{function of the \code{by} column that creates a new character vector of equal length (default = \code{fun})}

\item{fun.tre}{function of the tre tip labels that creates a new character vector of equal length (default = \code{fun})}
}
\value{
treedata object with an updated 'data' slot
}
\description{
Update table data based on string matches
}
