% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{add_tip_data_by_strain_name}
\alias{add_tip_data_by_strain_name}
\title{Extract tip data from a table based on strain name}
\usage{
add_tip_data_by_strain_name(tre, dat, by = "Strain")
}
\arguments{
\item{tre}{treedata object}

\item{dat}{data.frame containing the column \code{by}.}

\item{by}{the name of the column containing strain names in \code{dat} table}
}
\value{
treedata object with new 'data' slot
}
\description{
If there is an existing 'data' slot in the \code{tre} object, then it will
be replaced. The tree is expected to have tip labels that contain, but may
not be equal to, the strain name. Strain names are extracted from the tip
labels using the \code{extract_strain_name} function. After the data are
added, the tip labels are set to the strain names and the original labels
are stored in a new 'header' column.
}
