% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{clean_data}
\alias{clean_data}
\title{Clean the dataset}
\usage{
clean_data(d, remove_mixed = TRUE)
}
\arguments{
\item{df}{Datatable containing swine surveillance information}
}
\value{
Return a data table with standardized clade names and filtered datasets
}
\description{
Clean the dataset
}
