% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{collapse_c4}
\alias{collapse_c4}
\title{Collapse Cluster IV into C-IV and C-IVA}
\usage{
collapse_c4(d)
}
\arguments{
\item{The}{input dataset}
}
\value{
Data suitable for basic_plot
}
\description{
Collapse Cluster IV into C-IV and C-IVA
}
