% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{collapse_gamma}
\alias{collapse_gamma}
\title{Collapse gamma.[123] and gamma2 into gamma}
\usage{
collapse_gamma(d)
}
\arguments{
\item{The}{input dataset}
}
\value{
Data suitable for basic_plot
}
\description{
Collapse gamma.[123] and gamma2 into gamma
}
