% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{collapse_n2}
\alias{collapse_n2}
\title{Collapse 2002A, 2002B, 1998A and 1998B clades into 2002 and 1998}
\usage{
collapse_n2(d)
}
\arguments{
\item{The}{input dataset}
}
\value{
Data suitable for basic_plot
}
\description{
Collapse 2002A, 2002B, 1998A and 1998B clades into 2002 and 1998
}
