% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{common_barplot}
\alias{common_barplot}
\title{Barchart comparing the most common HA/NA/Constellation triplet}
\usage{
common_barplot(
  d,
  floorDateBy = "month",
  xfontsize = 10,
  yfontsize = 10,
  limits = NULL,
  title1 = glue::glue("count by {floorDateBy}"),
  title2 = glue::glue("\% by {floorDateBy}")
)
}
\description{
Barchart comparing the most common HA/NA/Constellation triplet
}
