% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{countByTimeUnit}
\alias{countByTimeUnit}
\title{Count number of entries by time chunk}
\usage{
countByTimeUnit(
  df,
  col_name = "H1",
  Date = "Date",
  tunit = "month",
  minDate = NULL,
  maxDate = NULL
)
}
\arguments{
\item{df}{is dataframe containing a column of Date and clade to count by.}

\item{col_name}{by default is "H1" but can change it to "N2","Subtype","region", etc.}

\item{Date}{by default is "Date" but can change if column name is different.}

\item{tunit}{is "month", "quarter" or "year".}

\item{minDate}{is default NULL, but can expand the date range lubridate::as_date("2019-01-01")}

\item{maxDate}{is default NULL, but can expand the date range}
}
\value{
a datatable with columns Date, Clade, n (counts)
}
\description{
Count number of entries by time chunk
}
