% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{countplot}
\alias{countplot}
\title{Make a count plot from a column}
\usage{
countplot(df, x_string, titletext)
}
\arguments{
\item{df}{data.frame swine surveillance data (i.e., output or \code{load_current})}

\item{x_string}{The influenze segment that will be plotted [H1,H3,N1,N3,PB2,PB1,PA,NP,M,NS]}

\item{titletext}{The title of the plot}
}
\value{
ggplot object
}
\description{
Make a count plot from a column
}
