% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{date2quarter}
\alias{date2quarter}
\title{Return regular or federal quarters}
\usage{
date2quarter(dates, fed = FALSE)
}
\arguments{
\item{dates}{Vector of date values}

\item{fed}{True for returning federal quarters (19Q1 -> 19Q2)}
}
\value{
Return a vector of quarters from the dates YYQ1 - YYQ4
}
\description{
Return regular or federal quarters
}
