% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{date_to_quarter}
\alias{date_to_quarter}
\title{Convert from date to fiscal quarter}
\usage{
date_to_quarter(datey)
}
\arguments{
\item{datey}{date}
}
\value{
quarter (Q1, Q2, Q3, Q4)
}
\description{
Convert from date to fiscal quarter
}
