% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{dates_to_str}
\alias{dates_to_str}
\title{Prepare a text description of a date range}
\usage{
dates_to_str(dates)
}
\arguments{
\item{dates}{a vector of dates}
}
\description{
For example, "Jun 2019 to Jun 2020"
}
