% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{extract_strain_name}
\alias{extract_strain_name}
\title{Extract the strain name from a vector of strings}
\usage{
extract_strain_name(x)
}
\arguments{
\item{x}{character vector (e.g., of FASTA headers)}
}
\value{
character vector with the same length as x (with possible missing values)
}
\description{
We assume the strain name follows the pattern 'A/.../<year>' For example:
'A/swine/Oklahoma/A01785514/2018'. We further assume that there are no '|'
characters in the sequence (since these are used as separators in FASTA
headers).
}
