% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{facetMaps}
\alias{facetMaps}
\title{Make state map}
\usage{
facetMaps(df, segment, counts = FALSE, normalization = "clade-max")
}
\arguments{
\item{df}{the input surveyllance data}

\item{segment}{the column to facet by}
}
\description{
Make state map
}
