% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_palette}
\alias{get_palette}
\title{Return the segment palette for the plots}
\usage{
get_palette(seg = "H1")
}
\arguments{
\item{seg}{H1, H3, N1, N2, PB2, PB1, PA, NP, M, NS}
}
\value{
Return a named vector of hex colors
}
\description{
Return the segment palette for the plots
}
