% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{group_by_nested_clade}
\alias{group_by_nested_clade}
\title{Create a new column in the tree data that groups}
\usage{
group_by_nested_clade(tre, by, to, na.rm = TRUE)
}
\arguments{
\item{tre}{treedata object}

\item{by}{string a column in the treedata@data slot}

\item{to}{string the name of the new column that will be created}

\item{na.rm}{logical should NA's be ignored in the \code{by} column?}
}
\value{
tre treedata object
}
\description{
Create a new column in the tree data that groups
}
