% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hana_barplots}
\alias{hana_barplots}
\title{Barchart comparing the most common clades over time}
\usage{
hana_barplots(
  d,
  floorDateBy = "month",
  xfontsize = 10,
  yfontsize = 10,
  limits = NULL,
  title1 = glue::glue("HA/NA phylogenetic-clade count by {floorDateBy}"),
  title2 = glue::glue("HA/NA phylogenetic-clade \% by {floorDateBy}")
)
}
\description{
The input is the full dataset which is then reduced to HA/NA counts before building.
}
