% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_hana_diff}
\alias{heatmap_hana_diff}
\title{Make a HANA diff heatmap}
\usage{
heatmap_hana_diff(d, cquarters, pquarters, font_size = 3, totals = FALSE)
}
\description{
The input is the full dataset which is then reduced to HA/NA counts before building.
}
