% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{increment_quarter}
\alias{increment_quarter}
\title{Increment fiscal quarter}
\usage{
increment_quarter(quarter)
}
\arguments{
\item{quarter}{The fiscal quarter (e.g., 20Q1)}
}
\value{
the next quarter (e.g., 20Q1 -> 20Q2 and 20Q4 -> 21Q1)
}
\description{
Increment fiscal quarter
}
