% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{labelFromQuarters}
\alias{labelFromQuarters}
\title{Make a label stating the quarterly range}
\usage{
labelFromQuarters(quarters)
}
\arguments{
\item{quarters}{a list of quarters}
}
\value{
A string describing the quarterly range (e.g., "[20Q1 to 21Q4]")
}
\description{
Make a label stating the quarterly range
}
