% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{load_file}
\alias{load_file}
\title{Load a new swine surveillance report}
\usage{
load_file(filename)
}
\arguments{
\item{filename}{The name of the input TAB-delimited file (the output of
`octofludb make masterlist`)}
}
\description{
Load a new swine surveillance report
}
