% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{load_segment_tree}
\alias{load_segment_tree}
\title{Load a tree with segment info}
\usage{
load_segment_tree(
  segment,
  filename = NULL,
  reference_data = NULL,
  fun.tre = function(x) {
     x <- sub("REF|", "", x)
     sub("([^|]+\\\\|[^|]+).*",
    "\\\\1", x, perl = TRUE)
 },
  fun.dat = function(gbid, strain) {
     paste0(gbid, "|", strain)
 }
)
}
\description{
Load a tree with segment info
}
