% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{orderConstellations}
\alias{orderConstellations}
\title{Make cannonical constellation ordering}
\usage{
orderConstellations(d)
}
\arguments{
\item{d}{A dataframe with a Constellation column}
}
\value{
The dataframe with the ordered Constellation column
}
\description{
Make cannonical constellation ordering
}
