% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_basic}
\alias{plot_basic}
\title{Stacked bar chart for phylocluster by month, quarter or year}
\usage{
plot_basic(
  d,
  segment = "H1",
  floorDateBy = "month",
  xfontsize = 10,
  yfontsize = 10,
  limits = NULL,
  title1 = glue::glue("{segment} phylogenetic-clade count by {floorDateBy}"),
  title2 = glue::glue("{segment} phylogenetic-clade \% by {floorDateBy}")
)
}
\arguments{
\item{d}{data.frame swine surveillance data (i.e., output or \code{load_current})}

\item{segment}{The influenze segment that will be plotted [H1,H3,N1,N3,PB2,PB1,PA,NP,M,NS]}

\item{floorDateBy}{An argument that is passed to
\code{lubridate::floor_date} specifying how to group dates (e.g., "month",
"week", "3 months", "year")}
}
\value{
ggplot object
}
\description{
Stacked bar chart for phylocluster by month, quarter or year
}
