% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_constellation}
\alias{plot_constellation}
\title{Make a Gene Constellation Heatmap of the data}
\usage{
plot_constellation(d)
}
\arguments{
\item{d}{data.frame swine surveillance data (i.e., output or \code{load_current})}
}
\value{
ggplot object
}
\description{
Make a Gene Constellation Heatmap of the data
}
