% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Make a HA NA Heatmap of the data}
\usage{
plot_heatmap(d)
}
\arguments{
\item{d}{data.frame swine surveillance data (i.e., output or \code{load_current})}
}
\value{
ggplot object
}
\description{
Make a HA NA Heatmap of the data
}
