% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{quadcountplot}
\alias{quadcountplot}
\title{Bar charts for H1, H3, N1, N2, depends on countplot function}
\usage{
quadcountplot(df, timespan_str)
}
\arguments{
\item{df}{data.frame swine surveillance data (i.e., output or \code{load_current})}

\item{timespan_str}{The title of the plot}
}
\value{
ggplot object
}
\description{
Bar charts for H1, H3, N1, N2, depends on countplot function
}
