% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{shared_legend_plot}
\alias{shared_legend_plot}
\title{Make two plots on top of eachother with a shared legend beneath}
\usage{
shared_legend_plot(p1, p2)
}
\arguments{
\item{p1}{Top plot (the shared legend is taken from this plot)}

\item{p2}{Bottom plot}
}
\value{
a ggplot2 object
}
\description{
This layout works well for wide barcharts. Both plots are assumed to have
the same legend, but this is not checked.
}
