% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{triple_barplots}
\alias{triple_barplots}
\title{Barchart comparing the most common HA/NA/Constellation triplet}
\usage{
triple_barplots(d, floorDateBy = "month", ...)
}
\description{
Barchart comparing the most common HA/NA/Constellation triplet
}
