% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barchart_bytime}
\alias{barchart_bytime}
\title{Plot barchart}
\usage{
barchart_bytime(
  df,
  value = "n",
  variable = "H1",
  palette = octoflushow::get_palette(variable),
  title = "",
  bartype = "stack",
  limits = NULL,
  tunit = "month",
  fed = FALSE,
  xfontsize = 10,
  yfontsize = 10
)
}
\arguments{
\item{df}{is the dataframe of clades and counts by date, an count_bytime output}

\item{value}{contains the counts}

\item{variable}{contains the clade names}

\item{palette}{is the named array of hex colors by clade}

\item{title}{is the title of the plot}

\item{bartype}{is either "stack" or "fill"}

\item{fed}{is either T or F if should use federal quarters}
}
\description{
Plot barchart
}
