% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hana_barplots}
\alias{hana_barplots}
\title{Barchart comparing the most common clades over time}
\usage{
hana_barplots(d, floorDateBy = "month", global = FALSE, ...)
}
\description{
Barchart comparing the most common clades over time
}
