% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{load_file}
\alias{load_file}
\title{Load a new swine surveillance report}
\usage{
load_file(filename, sheet = 1)
}
\arguments{
\item{filename}{The name of an excel file}

\item{sheet}{The sheet in the excel file that contains the data.}
}
\description{
Load a new swine surveillance report
}
