% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{plot_basic}
\alias{plot_basic}
\title{Make a basic barplot of the data}
\usage{
plot_basic(d)
}
\arguments{
\item{d}{data.frame swine surveillance data (i.e., output or \code{load_current})}
}
\value{
ggplot object
}
\description{
Make a basic barplot of the data
}
